Edit logic in index.js. Delete card template.
